Move gt-server implicits out to separate subproject
So that stac can just contain the datamodel and json serde
